Settings

Settings

AsyPlots.asy — Function.

asy()

Set the backend to Asymptote

source

AsyPlots.cairo — Function.

cairo()

Set the backend to Cairo

source

AsyPlots.junosvg — Function.

junosvg()

Set the default output format for the Juno plot pane to svg

source

AsyPlots.junopng — Function.

junopng()

Set the default output format for the Juno plot pane to png

source

AsyPlots.fitwidth — Function.

fitwidth()

Set the figures appearing in the Juno plot pane to match its width

source

AsyPlots.fitheight — Function.

fitheight()

Set the figures appearing in the Juno plot pane to match its height

source